Reset the level when you restart it.
I've separated the initialisation code from the creation code to do this. Also
restructured things a bit: now there's a separate Scene per level (or would be
if we had more than one level) and we look up the references to objects inside
it when we switch to the level.