Better HUD using canvas2d text rendering
Drastic changes to the HUD, looks a lot better now. Also added placeholder
icons for gold and life, no icon yet for the timer though.

Changed the tile size to 6 to make the levels a bit easier to negotiate. Some
of the LD commenters had said that it was too hard to find the exit.